added factories for mocked values
