+++ components/WarningNote/WarningNotes.spec.tsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import * as warningNotes from 'utils/api/warningNotes';
 import WarningNotes from './WarningNotes';
 
-import { mockedWarningNote } from 'fixtures/warningNotes.fixtures';
+import { mockedWarningNote } from 'factories/warningNotes';
 
 jest.mock('next/router', () => ({
   useRouter: () => ({
