--- fixtures/allocatedWorkers.fixtures.ts
@@ -1,17 +1,0 @@
-import { Allocation } from 'types';
-
-export const mockedAllocations: Allocation[] = [
-  {
-    allocatedWorkerTeam: 'Safeguarding and Reviewing Team',
-    allocatedWorker: 'Officer Name',
-    allocationEndDate: '2019-03-28 00:00:00',
-    allocationStartDate: '2019-03-28 00:00:00',
-    workerType: 'Consultant Social Worker',
-    id: 123,
-    personId: 1234,
-    caseStatus: 'Open',
-    personName: 'foo',
-    personAddress: 'the address',
-    personDateOfBirth: '2020-03-20',
-  },
-];
