--- fixtures/warningNotes.fixtures.ts
@@ -1,36 +1,0 @@
-import { WarningNote } from 'types';
-
-export const mockedWarningNote: WarningNote[] = [
-  {
-    id: 123,
-    type: 'Risk to Adults',
-    createdDate: new Date(2020, 11, 12),
-    createdBy: 'Foo',
-    reviewedDate: new Date(2020, 11, 13),
-    reviewedBy: 'Bar',
-    disclosedDate: new Date(2020, 11, 12),
-    disclosedHow: ['written'],
-    disclosedWithIndividual: 'Yes',
-    disclosedDetails: 'lorem ipsum',
-    nextReviewDate: new Date(2020, 11, 31),
-    notes: 'i am a note',
-    discussedWithManager: 'the big boss',
-    discussedWithManagerDate: new Date(2020, 11, 12),
-    status: 'open',
-    reviews: [],
-  },
-  {
-    id: 234,
-    type: 'Risk to Staff',
-    createdDate: new Date(2020, 11, 22),
-    createdBy: 'Foo',
-    disclosedWithIndividual: 'No',
-    disclosedDetails: 'nope',
-    discussedWithManager: 'the big boss',
-    discussedWithManagerDate: new Date(2020, 11, 12),
-    nextReviewDate: new Date(2021, 1, 1),
-    notes: 'a lot to talk about',
-    status: 'open',
-    reviews: [],
-  },
-];
