+++ package.json
@@ -77,6 +77,7 @@
     "eslint-plugin-jest": "^24.1.3",
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-react": "^7.22.0",
+    "fishery": "^1.2.0",
     "http-proxy-middleware": "^1.0.6",
     "husky": "^4.3.7",
     "identity-obj-proxy": "^3.0.0",
